#61 [Shaz] Restructured health endpoint
